+++ examples/bad/test_aaa03_04.py
@@ -1,0 +1,4 @@
+def test():
+    x = 1
+    result = x**2
+    assert result == 4