+++ tests/function/test_init.py
@@ -23,7 +23,6 @@ def test(first_node_with_tokens, lines):
         '    pass\n',
     ]
     assert result.act_node is None
-    assert result._errors is None
     assert result.line_markers == [LineType.unprocessed, LineType.unprocessed]
     assert result.line_markers.fn_offset == 3
     assert result.first_line_no == 3